@@ -2673,7 +2673,12 @@ namespace PInvoke
 
         [DllImport(nameof(Kernel32), CharSet = CharSet.Unicode, SetLastError = true)]
         [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool ReadConsoleOutput(IntPtr hConsoleOutput, out CHAR_INFO lpBuffer, COORD dwBufferSize, COORD dwBufferCoord, ref SMALL_RECT lpReadRegion);
+        public static extern unsafe bool ReadConsoleOutput(
+            IntPtr hConsoleOutput,
+            [Friendly(FriendlyFlags.Out | FriendlyFlags.Array)] CHAR_INFO* lpBuffer,
+            COORD dwBufferSize,
+            COORD dwBufferCoord,
+            ref SMALL_RECT lpReadRegion);
 
         /// <summary>
         /// Reads character input from the console input buffer and removes it from the buffer.
