@@ -13,7 +13,7 @@ namespace PInvoke
         [NoFriendlyOverloads]
         public static unsafe NTSTATUS RtlGetVersion(Kernel32.OSVERSIONINFOEX* versionInformation) => RtlGetVersion((Kernel32.OSVERSIONINFO*)versionInformation);
 
-        /// <inheritdoc cref="RtlGetVersion(ref Kernel32.OSVERSIONINFO)"/>
+        /// <inheritdoc cref="RtlGetVersion(Kernel32.OSVERSIONINFOEX*)"/>
         public static unsafe NTSTATUS RtlGetVersion(ref Kernel32.OSVERSIONINFOEX versionInformation)
         {
             fixed (Kernel32.OSVERSIONINFOEX* versionInformationLocal = &versionInformation)
